llama_index_example/pyproject.toml
[project]
name = "llama-index-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "create-agent-app",
    "llama-index>=0.10.0",
    "langwatch>=0.2.9",
    "langwatch-scenario>=0.1.3",
    "openinference-instrumentation-llama-index>=0.1.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
create-agent-app = { path = "../", editable = true } 